requirements.txt
requests
snowflake-connector-python
python-dotenv
pandas
spotipy
cryptography
